Implemented Apollo local state for Cart
- Used reactive variables
- Syncs localstorage data with useQuery on initial load to fill cart
- Sync cart with server data on demand
