+++ client/src/pages/Product/Product.tsx
@@ -1,43 +1,25 @@
 import { useQuery } from "@apollo/client";
-import { ChangeEvent, FormEvent, useState } from "react";
-import { Button, Col, Container, Form, Row } from "react-bootstrap";
+import { useState } from "react";
+import { Col, Container, Row } from "react-bootstrap";
 import { Link, useParams } from "react-router-dom";
 import Cart from "../../components/Cart";
-import useCartStore from "../../context/useStore";
 import { GetSingleProductQuery } from "../../generated/graphql";
 import { GET_SINGLE_PRODUCT } from "../../queries/Product";
-import { toPriceNumber } from "../../utils/numbers";
+import AddProductForm from "./AddProductForm";
 import RelatedProducts from "./RelatedProducts";
 
 const Product = () => {
     const { id } = useParams();
-    const addProductToCart = useCartStore((state) => state.addProductToCart);
     const [selectedImage, setSelectedImage] = useState(0);
     const [showCart, setShowCart] = useState(false);
-    const [amount, setAmount] = useState(1);
     const { data, loading, error } = useQuery<GetSingleProductQuery>(
         GET_SINGLE_PRODUCT,
         { variables: { id: +id! } }
     );
 
-    const handleAmount = (e: ChangeEvent<HTMLInputElement>) => {
-        setAmount(+e.target.value);
-    };
-
     // Cart Modal
     const handleShowCart = () => setShowCart(true);
     const handleCloseCart = () => setShowCart(false);
-
-    const handleSubmit = (e: FormEvent) => {
-        e.preventDefault();
-        // add to cart variable
-
-        if (data?.product && data.product.inventory >= amount) {
-            addProductToCart({ product: data.product, amount: amount });
-        }
-        // open cart modal
-        handleShowCart();
-    };
 
     if (error || data?.product === null) {
         return (
@@ -164,62 +146,10 @@ const Product = () => {
                                     </div>
                                 </>
                             )}
-                            <Form onSubmit={handleSubmit}>
-                                <Row className="d-flex justify-content-between align-items-center gap-3 pt-3 mb-4">
-                                    {data.product.inventory > 0 ? (
-                                        <Col>
-                                            <Form.Control
-                                                className="w-25"
-                                                type="number"
-                                                value={amount}
-                                                min={1}
-                                                max={data.product.inventory}
-                                                onChange={handleAmount}
-                                            />
-                                        </Col>
-                                    ) : (
-                                        <Col className="text-danger fs-4">
-                                            Out of stock
-                                        </Col>
-                                    )}
-
-                                    <Col className="fs-3">
-                                        <div className="d-flex gap-3 align-items-center">
-                                            <div>
-                                                {!!data.product.discount && (
-                                                    <s className="text-muted fs-5">{`${toPriceNumber(
-                                                        amount *
-                                                            data.product.price
-                                                    )} $`}</s>
-                                                )}
-                                            </div>
-                                            <div
-                                                className={
-                                                    data.product.discount
-                                                        ? "text-danger fs-3 lh-1"
-                                                        : "fs-3 lh-1"
-                                                }
-                                            >
-                                                {`${toPriceNumber(
-                                                    ((100 -
-                                                        data.product.discount) /
-                                                        100) *
-                                                        amount *
-                                                        data.product.price
-                                                )} $`}
-                                            </div>
-                                        </div>
-                                    </Col>
-                                </Row>
-                                <Button
-                                    type="submit"
-                                    variant="success"
-                                    disabled={!data.product.inventory}
-                                    style={{ width: "10rem" }}
-                                >
-                                    Add to Cart
-                                </Button>
-                            </Form>
+                            <AddProductForm
+                                data={data}
+                                handleShowCart={handleShowCart}
+                            />
                         </Col>
                     </>
                 )}
